regroup handler with main page
main page was empty anyway and I wanted to to put Process btn on the
right of the title as for the other pages
