--- src/main/webapp/app/spreadsheet/spreadsheet-handler/spreadsheet-handler.component.html
@@ -1,11 +1,0 @@
-<div>
-    <jhi-dragndrop fileLimit="1" (filesChange)="validateWorkbook($event)"></jhi-dragndrop>
-    <div class="btn-group flex-btn-group-container">
-        <button [disabled]="workbookValidation === undefined || !workbookValidation.valid"
-                (click)="processWB()"
-                class="btn btn-danger">
-            <span class="fa fa-pencil"></span>
-            <span class="hidden-md-down" jhiTranslate="validation.handler.process-btn"></span>
-        </button>
-    </div>
-</div>
