+++ src/main/webapp/app/spreadsheet/spreadsheet.component.html
@@ -1,8 +1,15 @@
 <div>
     <h2>
         <span jhiTranslate="validation.title">Learners</span>
+        <button [disabled]="workbookValidation === undefined || !workbookValidation.valid"
+                (click)="processWB()"
+                class="btn btn-danger float-right">
+            <span class="fa fa-pencil"></span>
+            <span jhiTranslate="validation.handler.process-btn"></span>
+        </button>
     </h2>
     <jhi-alert></jhi-alert>
-    <jhi-spreadsheet-handler resource="registration"></jhi-spreadsheet-handler>
-    <jhi-spreadsheet-validation></jhi-spreadsheet-validation>
+    <jhi-dragndrop fileLimit="1" (filesChange)="validateWorkbook($event)"></jhi-dragndrop>
+    <br/>
+    <jhi-spreadsheet-validation *ngIf="workbookValidation !== undefined && !workbookValidation.valid"></jhi-spreadsheet-validation>
 </div>
