+++ src/main/webapp/app/spreadsheet/spreadsheet.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SpreadsheetComponent } from './spreadsheet.component';
-import { SpreadsheetHandlerComponent } from './spreadsheet-handler/spreadsheet-handler.component';
 import { DragndropModule } from '../dragndrop/dragndrop.module';
 import { SpreadsheetService } from './spreadsheet.service';
 import { SpreadsheetValidationComponent } from './spreadsheet-validation/spreadsheet-validation.component';
@@ -21,7 +20,6 @@ import { ClickandrungwSharedCommonModule } from '../shared';
     ],
     declarations: [
         SpreadsheetComponent,
-        SpreadsheetHandlerComponent,
         SpreadsheetValidationComponent,
         SheetValidationComponent,
         TableValidationComponent
